+++ src/__tests__/api.test.ts
@@ -30,7 +30,7 @@ jest.mock("https", () => {
             agentInstance.passphrase = opts.passphrase;
           }
           return agentInstance;
-        }
+        },
       ),
   };
 });