+++ src/record/delete/index.ts
@@ -8,6 +8,7 @@ import { deleteByRecordNumber } from "./usecases/deleteByRecordNumber";
 import { logger } from "../../utils/log";
 import type { SupportedImportEncoding } from "../../utils/file";
 import { readFile } from "../../utils/file";
+import { isMismatchEncoding } from "../../utils/encoding";
 import { parseRecords } from "./parsers";
 import { RunError } from "../error";
 
@@ -18,7 +19,7 @@ export type Options = {
 };
 
 export const run: (
-  argv: RestAPIClientOptions & Options
+  argv: RestAPIClientOptions & Options,
 ) => Promise<void> = async (options) => {
   try {
     const { app, filePath, encoding, ...restApiClientOptions } = options;
@@ -39,13 +40,17 @@ const deleteRecordsByFile = async (
   apiClient: KintoneRestAPIClient,
   app: string,
   filePath: string,
-  encoding?: SupportedImportEncoding
+  encoding?: SupportedImportEncoding,
 ): Promise<void> => {
+  if (encoding) {
+    await validateEncoding(filePath, encoding);
+  }
+
   const recordNumbers = await getRecordNumbersFromFile(
     apiClient,
     app,
     filePath,
-    encoding
+    encoding,
   );
   if (recordNumbers.length === 0) {
     logger.warn("The specified CSV file does not have any records.");
@@ -59,7 +64,7 @@ const getRecordNumbersFromFile = async (
   apiClient: KintoneRestAPIClient,
   app: string,
   filePath: string,
-  encoding?: SupportedImportEncoding
+  encoding?: SupportedImportEncoding,
 ): Promise<RecordNumber[]> => {
   const fieldsJson = await apiClient.app.getFormFields({ app });
   const recordNumberFieldCode = getRecordNumberFieldCode(fieldsJson.properties);
@@ -73,7 +78,7 @@ const getRecordNumbersFromFile = async (
 };
 
 const getRecordNumberFieldCode = (
-  properties: FieldsJson["properties"]
+  properties: FieldsJson["properties"],
 ): string => {
   let recordNumberFieldCode = "";
   for (const property of Object.values(properties)) {
@@ -84,4 +89,15 @@ const getRecordNumberFieldCode = (
   }
 
   return recordNumberFieldCode;
+};
+
+const validateEncoding: (
+  filePath: string,
+  encoding: SupportedImportEncoding,
+) => Promise<void> = async (filePath, encoding) => {
+  if (await isMismatchEncoding(filePath, encoding)) {
+    throw new Error(
+      `Failed to decode the specified CSV file.\nThe specified encoding (${encoding}) might mismatch the actual encoding of the CSV file.`,
+    );
+  }
 };