+++ src/record/export/schema/__tests__/index.test.ts
@@ -18,7 +18,7 @@ describe("createSchema", () => {
   const patterns = [formLayout, formLayoutNoSystemFields, userSelected];
   it.each(patterns)("$description", (pattern) => {
     expect(createSchema(pattern.input, pattern.transformer)).toStrictEqual(
-      pattern.expected
+      pattern.expected,
     );
   });
 });