+++ src/record/import/repositories/parsers/__tests__/error.test.ts
@@ -9,11 +9,11 @@ describe("ParserError", () => {
   it("should be constructed with CsvError", () => {
     const csvError = new CsvError(
       "CSV_QUOTE_NOT_CLOSED",
-      "Quote Not Closed: the parsing is finished with an opening quote at line 2"
+      "Quote Not Closed: the parsing is finished with an opening quote at line 2",
     );
     const error = new ParserError(csvError);
     expect(error.toString()).toBe(
-      "Failed to parse input\nCSV_QUOTE_NOT_CLOSED: Quote Not Closed: the parsing is finished with an opening quote at line 2\n"
+      "Failed to parse input\nCSV_QUOTE_NOT_CLOSED: Quote Not Closed: the parsing is finished with an opening quote at line 2\n",
     );
   });
 });